Roles and Capabilities'
Flesh out previous code a lot more.
The first thing to be guarded by a capability is the roles editor itself.
Also added a load of unit tests.
Adding the unit tests entailed also fixing some issues with unit tests, and adding the ability for a unit test to click inside a popup.