triggerOnGlobalWrite lets a derived view update when a global changes.
Right now this only triggers when a client updates a global, or if another server module explicitly calls to run triggers after updating a global.

In the future we may change this so that all global writes in server store run triggers. I'm just being a bit cautious to begin with.